+++ Screens/CreateUserScreen.py
@@ -9,7 +9,7 @@ from tkinter_objects.ControllerMainObject import *
 from TableObjects import *
 
 class CreateUser(ChildFrame):
-    def __init__(self, master, controller, admin_mode, create_mode, **kwargs) -> None:
+    def __init__(self, master, controller, create_mode, **kwargs) -> None:
         super().__init__(master, controller, **kwargs)
 
         label_text = ""
@@ -17,24 +17,23 @@ class CreateUser(ChildFrame):
             label_text += "Create"
         else:
             label_text += "Update"
-        if admin_mode:
-            label_text += " Admin"
-        else:
-            label_text += " User"
-        tk.Label(self, text=label_text).grid(row=0, column=0, columnspan=2)
+        tk.Label(self, text=label_text + "user").grid(row=0, column=0, columnspan=2)
+
+        self.admin_check_button = tk.Checkbutton(self, text="User is Admin")
+        self.admin_check_button.grid(row=1, column=0)
         
-        tk.Label(self, text="Username").grid(row=1, column=0)
+        tk.Label(self, text="Username").grid(row=2, column=0)
         self.username = tk.StringVar()
         self.username_entry = tk.Entry(self, textvariable=self.username)
-        self.username_entry.grid(row=1, column=1)
+        self.username_entry.grid(row=2, column=1)
 
-        tk.Label(self, text="Password").grid(row=2, column=0)
+        tk.Label(self, text="Password").grid(row=3, column=0)
         self.password = tk.StringVar()
         self.password_entry = tk.Entry(self, textvariable=self.password)
-        self.password_entry.grid(row=2, column=1)
+        self.password_entry.grid(row=3, column=1)
 
         self.control_frame = tk.Frame(self)
-        self.control_frame.grid(row=3, column=0, columnspan=2)
+        self.control_frame.grid(row=4, column=0, columnspan=2)
 
         
         button_text = ""